Add classification stage; add time columns

+++ preprocess_gisaid.py
@@ -1,4 +1,5 @@
 import argparse
+import datetime
 import json
 import logging
 import os
@@ -25,13 +26,19 @@ def ln_sf(source, target):
     os.symlink(source, target)
 
 
+DATE_FORMATS = {4: "%Y", 7: "%Y-%m", 10: "%Y-%m-%d"}
+
+
+def parse_date(string):
+    return datetime.datetime.strptime(string, DATE_FORMATS[len(string)])
+
+
 POOL = None
 
 
-def pmap(fn, args):
+def parallel_map(fn, args):
     # Avoid multiprocessing when running under pdb.
-    main_module = sys.modules["__main__"]
-    if not hasattr(main_module, "__spec__"):
+    if not hasattr(sys.modules["__main__"], "__spec__"):
         return [fn(*a) for a in args]
 
     global POOL
@@ -40,20 +47,25 @@ def pmap(fn, args):
     return POOL.starmap(fn, args)
 
 
-def update_shards(shard_names):
-    infile = os.path.expanduser("~/data/gisaid/provision.json")
-    logger.info(f"Splitting {infile} into {args.num_shards} shards")
-    if not os.path.exists(infile):
-        raise OSError("Each user must independently request a data feed from gisaid.org")
-    with ExitStack() as stack:
-        f = stack.enter_context(open(infile))
-        shards = [stack.enter_context(open(shard_name, "w"))
-                  for shard_name in shard_names]
-        for i, line in enumerate(f):
-            shards[i % args.num_shards].write(line)
-            if i % args.log_every == 0:
-                print(".", end="", flush=True)
-    logger.info(f"split {i + 1} lines")
+def update_shards(args):
+    shard_names = [f"results/gisaid.{i:03d}-of-{args.num_shards:03d}.json"
+                   for i in range(args.num_shards)]
+    if args.force or not all(map(os.path.exists, shard_names)):
+        args.force = True
+        infile = os.path.expanduser("~/data/gisaid/provision.json")
+        logger.info(f"Splitting {infile} into {args.num_shards} shards")
+        if not os.path.exists(infile):
+            raise OSError("Each user must independently request a data feed from gisaid.org")
+        with ExitStack() as stack:
+            f = stack.enter_context(open(infile))
+            shards = [stack.enter_context(open(shard_name, "w"))
+                      for shard_name in shard_names]
+            for i, line in enumerate(f):
+                shards[i % args.num_shards].write(line)
+                if i % args.log_every == 0:
+                    print(".", end="", flush=True)
+        logger.info(f"split {i + 1} lines")
+    return shard_names
 
 
 STATS = ["date", "location", "length", "nchars"]
@@ -78,8 +90,11 @@ def get_stats(args, shard_names):
     if args.force or not os.path.exists(cache_file):
         logger.info("Computing statistics")
         stats = {key: Counter() for key in STATS}
-        for result in pmap(_get_stats, [(args, s) for s in shard_names]):
-            for key, value in result.items():
+        shards = parallel_map(_get_stats, [
+            (args, s) for s in shard_names
+        ])
+        for shard in shards:
+            for key, value in shard.items():
                 stats[key].update(value)
         logger.info(f"saving {cache_file}")
         with open(cache_file, "wb") as f:
@@ -95,15 +110,25 @@ def get_stats(args, shard_names):
 
 def _make_sketch(args, sketcher, shard_name):
     sequences = []
+    fields = ["accession_id", "collection_date", "location", "add_location"]
+    columns = {key: [] for key in fields}
+    columns["day"] = []
     with open(shard_name) as f:
         for i, line in enumerate(f):
             datum = json.loads(line)
+            if len(datum["covv_collection_date"]) < 7:
+                continue  # Drop rows with no month information.
+            date = parse_date(datum["covv_collection_date"])
+            if date < args.start_date:
+                continue  # Drop rows before start date.
             seq = datum["sequence"].replace("\n", "")
             parts = re.findall("[ACGT]+", seq)
-            if args.min_nchars <= sum(map(len, parts)) <= args.max_nchars:
-                sequences.append(parts)
-            if i + 1 == args.truncate:
-                break
+            if not (args.min_nchars <= sum(map(len, parts)) <= args.max_nchars):
+                continue  # Drop rows with too few nucleotides.
+            sequences.append(parts)
+            for key in fields:
+                columns[key].append(datum["covv_" + key])
+            columns["day"].append((date - args.start_date).days)
 
     sketch = sketcher.init_sketch(len(sequences))
     for i, parts in enumerate(sequences):
@@ -111,72 +136,103 @@ def _make_sketch(args, sketcher, shard_name):
         if i % args.log_every == 0:
             print(".", end="", flush=True)
 
-    return sketch
+    return sketch, columns
 
 
 def make_sketch(args, shard_names):
-    cache_file = f"results/gisaid.sketch.{args.k}.pt"
+    cache_file = f"results/gisaid.sketch.{args.k}.{args.num_clocks}.pt"
     sketcher = ClockSketcher(k=args.k, num_clocks=args.num_clocks)
     if args.force or not os.path.exists(cache_file):
         args.force = True
-        logger.info("Sketching k-mers bags")
-        sketches = pmap(_make_sketch, [(args, sketcher, s) for s in shard_names])
-        clocks = torch.cat([s.clocks for s in sketches])
-        count = torch.cat([s.count for s in sketches])
+        logger.info("Sketching k-mer bags")
+        shards = parallel_map(_make_sketch, [
+            (args, sketcher, s) for s in shard_names
+        ])
+        clocks = torch.cat([s.clocks for s, _ in shards])
+        count = torch.cat([s.count for s, _ in shards])
         sketch = ClockSketch(clocks, count)
-        # TODO
-        # logger.info("greedily finding clusters")
+        columns = {}
+        for _, s in shards:
+            for key, column in s.items():
+                columns.setdefault(key, []).extend(column)
         result = {
-            "k": args.k,
-            "num_clocks": args.num_clocks,
+            "args": args,
             "sketch": sketch,
+            "columns": columns,
         }
         torch.save(result, cache_file)
         logger.info(f"saving {cache_file}")
     else:
         result = torch.load(cache_file)
         sketch = result["sketch"]
+        columns = result["columns"]
     ln_sf(cache_file, "results/gisaid.sketch.pt")
-    return sketcher, sketch
+    return sketcher, sketch, columns
 
 
 def cluster(args, sketcher, sketch):
-    cache_file = f"results/gisaid.cluster.{args.num_clusters}.{args.cluster_radius}.{args.cluster_epochs}.pt"
+    cache_file = (f"results/gisaid.cluster.{args.cluster_radius}.{args.max_clusters}."
+                  f"{args.min_cluster_size}.{args.cluster_epochs}.pt")
     if args.force or not os.path.exists(cache_file):
         args.force = True
-        logger.info(f"Clustering {len(sketch)} taxa into {args.num_clusters} clusters")
-        result = sketcher.find_clusters(sketch, num_clusters=args.num_clusters,
-                                        radius=args.cluster_radius,
-                                        epochs=args.cluster_epochs)
+        logger.info(f"Clustering {len(sketch)} taxa into {args.max_clusters} clusters")
+        full_clusters, full_weights = sketcher.find_clusters(
+            sketch,
+            num_clusters=args.max_clusters,
+            radius=args.cluster_radius,
+            epochs=args.cluster_epochs,
+        )
+
+        keep = full_weights >= args.min_cluster_size
+        logger.info(f"keeping {keep.sum():d}/{args.max_clusters} largest clusters")
+        weights = full_weights[keep]
+        clusters = full_clusters[keep]
+
+        logger.info(f"classifying {len(sketch)} sketches against {len(clusters)} clusters")
+        class_probs = sketcher.classify(sketch, clusters, weights,
+                                        radius=args.cluster_radius)
+        result = {
+            "args": args,
+            "full_clusters": full_clusters,
+            "full_weights": full_weights,
+            "clusters": clusters,
+            "weights": weights,
+            "class_probs": class_probs,
+        }
         torch.save(result, cache_file)
     else:
         result = torch.load(cache_file)
     ln_sf(cache_file, "results/gisaid.cluster.pt")
+    class_probs = result["class_probs"]
+    perplexity = (class_probs.log().neg_()
+                             .clamp_(max=torch.finfo(class_probs.dtype).max)
+                             .mul_(class_probs).sum(-1).exp())
+    logging.info(f"class perplexity min={perplexity.min():0.2f}, "
+                 f"median={perplexity.median():0.1f}, "
+                 f"mean={perplexity.mean():0.1f}, "
+                 f"max={perplexity.max():0.1f}")
     return result
 
 
 def main(args):
-    shard_names = [f"results/gisaid.{i:03d}-of-{args.num_shards:03d}.json"
-                   for i in range(args.num_shards)]
-    if args.force or not all(map(os.path.exists, shard_names)):
-        args.force = True
-        update_shards(shard_names)
+    # Shard and summarize data.
+    shard_names = update_shards(args)
     stats = get_stats(args, shard_names)
 
-    # Drop low quality sequences.
+    # Configure dropping of low quality sequences.
     assert 0 <= args.min_nchars_rel <= 1 <= args.max_nchars_rel
     nchars = stats["nchars"]
     mean_nchars = sum(k * v for k, v in nchars.items()) / sum(nchars.values())
     args.max_nchars = int(args.max_nchars_rel * mean_nchars)
     args.min_nchars = int(args.min_nchars_rel * mean_nchars)
-    num_ok = sum(v for k, v in nchars.items()
-                 if args.min_nchars < k < args.max_nchars)
-    total = sum(nchars.values())
-    logger.info(f"Keeping {num_ok}/{total} = {100*num_ok/total:0.1f}% of sequences")
 
-    if args.truncate:
-        shard_names = shard_names[:1]
-    sketcher, sketch = make_sketch(args, shard_names)
+    # Make sketches.
+    sketcher, sketch, columns = make_sketch(args, shard_names)
+    kept = len(sketch)
+    total = sum(nchars.values())
+    logger.info(f"kept {kept}/{total} = {100*kept/total:0.1f}% of sequences")
+
+    # Cluster taxa.
     clustering = cluster(args, sketcher, sketch)
     assert clustering
 
@@ -185,17 +241,19 @@ if __name__ == "__main__":
     mp.set_start_method("spawn")
 
     parser = argparse.ArgumentParser(description="Preprocess GISAID data")
+    parser.add_argument("--start-date", default="2019-12-01")
     parser.add_argument("--min-nchars-rel", default=0.95, type=float)
     parser.add_argument("--max-nchars-rel", default=1.05, type=float)
     parser.add_argument("--k", default=20, type=int)
     parser.add_argument("--num-clocks", default=256, type=int)
-    parser.add_argument("--num-clusters", default=200, type=int)
-    parser.add_argument("--cluster-radius", default=100, type=int)
-    parser.add_argument("--cluster-epochs", default=2, type=int)
-    parser.add_argument("-s", "--num-shards", default=mp.cpu_count(), type=int)
+    parser.add_argument("--max-clusters", default=200, type=int)
+    parser.add_argument("--min-cluster-size", default=10, type=int)
+    parser.add_argument("--cluster-radius", default=40, type=int)
+    parser.add_argument("--cluster-epochs", default=10, type=int)
+    parser.add_argument("--num-shards", default=mp.cpu_count(), type=int)
     parser.add_argument("-l", "--log-every", default=1000, type=int)
     parser.add_argument("-f", "--force", action="store_true")
-    parser.add_argument("--truncate", default=0, type=int)
     args = parser.parse_args()
+    args.start_date = parse_date(args.start_date)
 
     main(args)
